Refactored the code to made the repository more leand and scope business logic in service layer.

+++ SubRedditAPI/Interfaces/IRateLimitService.cs
@@ -2,7 +2,7 @@
 {
     public interface IRateLimitService
     {
-        bool IsRequestAtRateLimit(string apiBeingCalled, int rateLimitRemaining);
+        bool IsRequestAtRateLimit(string apiBeingCalled, double rateLimitRemaining);
         object? ReturnCachedData(string cacheKey);
     }
 }
